+++ Assets/Package/Runtime/Scripts/ContainerModuleLoader.cs
@@ -64,11 +64,23 @@ namespace MrWatts.Internal.FuelInject
                 .As<IInjector<Scene>>()
                 .SingleInstance();
 
+            builder.RegisterType<ConfigurableComponentContextProxy>()
+                .AsSelf()
+                .As<IComponentContext>()
+                .SingleInstance();
+
             IContainer container = builder.Build();
+
+            /*
+                Proxy the container (IComponentContext) through a proxy class, because the IComponentContext may not be
+                bound directly using container.Register<IComponentContext>(context => context) as `context` may not be
+                stored.
+            */
+            container.Resolve<ConfigurableComponentContextProxy>().Delegatee = container;
 
             gameObject.AddComponent<UnityKernel>();
 
-            container.Resolve<IInjector<Scene>>().Inject(gameObject.scene, container);
+            container.Resolve<IInjector<Scene>>().Inject(gameObject.scene);
         }
 
         private void LoadModules(ContainerBuilder builder)